(a) rename Cacher to FunctionCache (b) fix importing global variables (c) remove write/read cache from FunctionCache (d) remove asynchronous elements (e) change name to id (f) rename clearMemory to clearCache
(a): "cacher" was too close to all the other cache-adjacent names in this project. Plus, I'm thinking the FunctionCache class will be just a part of this project (instead of the whole thing), and so it's name should denote that.
(b): This one stumped me for a while. But apparently whenever you export a variable it basically becomes a constant, whether it was declared that way or not. So I made one variable (globals) and I don't redefine it, I just change aspects of it.
(c): I sort of refactored the actual writing to the cache. So now cache.txt will be written to no more than once every 250 milliseconds (or whatever time you give it), instead of once every time the cacheObj is mutated.
(d): moving the actual writing/reading elsewhere allowed me to remove the asynchronous elements from this file.
(e): More apt. Because, really, this just needs to be something unique so that an instance of FunctionCache can find it's section of cacheObj reliably.
(f): This name is more apt, because it's not just mutating the instance of FunctionCache's memory, it's also mutating cacheObj.