(a) rename global-variables to globals (b) put all globals under one const
a: This name is shorter, and it's more consistent now that I'm doing everything through one variable called "globals".
b: The code was not working before because exported variables become constants, and I didn't know that. So I put all the global variables under this one object (called "globals") which can have it's properties mutated while still being a const.