(a) use globals instead of global-variables (b) remove temporary cacheObj variable (c) add startCaching function
a: I say more about why this change was neccesary in previous commits. Changing this fixed a bug.
b: This was never supposed to be in this file. It's was a temporary workaround to a problem that is now solved (see a)
c: Writing to the cache once every x milliseconds will save  computing power. Before, I was having it write to the cache every time cacheObj was mutated, which I realized was way more than neccesary.
There's also this business of cacheUpToDate. That's because I didn't like the idea that my computer would be sitting there writing and rewriting the cache even if I'm not doing anything. So I made this variable so that it only writes to the cache if something is changed.
HOWEVER: this means that every time I mutate cacheObj, I need to remember to set globals.cacheUpToDate to false, otherwise cache.txt may not be updated. (foreshadowing future bugs?)