delete FunctionCache.js
I sort of had an epiphany, and I was like "I know what I want this project to look like!" And so I decided that FunctionCache was sort of outside the scope of the project. I want the project to be the basic caching function, and then you can make your own function cache if you need to.