(a) delete startCaching (b) add requestWrite (c) add store and retrieve (d) reorder functions (e) add documentation
a/b: I was thinking about it and I just didn't like having that caching loop going in the background. It was stressing me out. It was like "will it keep looping after I close the program?" So I came up with this different way to do it.
c: honestly this kinda goes along with a/b; it's there to make things simpler for the end user.
d: just seemed like a better order.
e: I figured it'd be good to add documentation.